pyproject.toml
[tool.poetry]
name = "tarzan"
version = "0.0.0"
description = "Tarzan, the Django explorer"
authors = ["Bruno Peixoto <brunolnetto@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.7"
django-rest-swagger = "^2.2.0"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
drf-yasg = "^1.21.7"
django = "^4.2.4"
pytest = "^7.4.1"
pytest-django = "^4.5.2"
coverage = "^7.3.0"
ruff = "^0.0.287"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
